satellite-pipeline/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
geopandas>=0.10.0
shapely>=1.8.0
rasterio>=1.2.0
pyproj>=3.0.0

# Earth Engine dependencies
earthengine-api>=0.1.320
ee-extra>=0.0.15
geemap>=0.19.0

# Data processing and analysis
scikit-learn>=1.0.0
xarray>=0.20.0
dask>=2022.1.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0.0
python-dotenv>=0.19.0
click>=8.0.0

# Time-series specific
statsmodels>=0.13.0

# Visualization (for debugging)
matplotlib>=3.5.0
seaborn>=0.11.0

# API
fastapi>=0.78.0
uvicorn>=0.17.0

# Storage
zarr>=2.11.0
fsspec>=2022.1.0 